requirements.txt
email-validator
dnspython
streamlit
pandas
xlsxwriter
openpyxl
